combine NotebookShow container and display
