--- src/components/notebooks/notebook_show_container.jsx
@@ -1,16 +1,0 @@
-import { connect } from 'react-redux';
-import NotebookShow from './notebook_show';
-import { fetchSingleNotebook } from '../../actions/notebooks_actions';
-
-const mapStateToProps = ({ notebook }) => ({
-  notebook,
-});
-
-const mapDispatchToProps = dispatch => ({
-  fetchSingleNotebook: notebookId => dispatch(fetchSingleNotebook(notebookId)),
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(NotebookShow);
